Use JobConfig object rather than passing around arguments
